// apps/searchneu/app/graduate/common/types.ts
export enum NUPathEnum {
    ND = "Natural/Designed World",
    EI = "Creative Express/Innov",
    IC = "Interpreting Culture",
    FQ = "Formal/Quant Reasoning",
    SI = "Societies/Institutions",
    AD = "Analyzing/Using Data",
    DD = "Difference/Diversity",
    ER = "Ethical Reasoning",
    WF = "1st Yr Writing",
    WD = "Adv Writ Dscpl",
    WI = "Writing Intensive",
    EX = "Integration Experience",
    CE = "Capstone Experience",
  }
  
  /**
   * Represents one of the seasons in which a student can take a course, as
   * abbreviated by Northeastern.
   */
  export enum SeasonEnum {
    FL = "FL",
    SP = "SP",
    S1 = "S1",
    S2 = "S2",
    SM = "SM",
  }
  
  /** A Status is one of on CO-OP, CLASSES, or INACTIVE */
  export enum StatusEnum {
    COOP = "COOP",
    CLASSES = "CLASSES",
    INACTIVE = "INACTIVE",
    HOVERINACTIVE = "HOVERINACTIVE",
    HOVERCOOP = "HOVERCOOP",
  }
  
  export type Status = keyof typeof StatusEnum;
  export type Season = keyof typeof SeasonEnum;
  
  /** A SearchNEU prerequisite object. */
  export type INEUReq = INEUAndReq | INEUOrReq | INEUReqCourse;
  /**
   * A SearchNEU AND prerequisite object.
   *
   * @param type   The type of the SearchNEU prerequisite.
   * @param values The prerequisites that must be completed for this prereq. to be
   *   marked as done.
   */
  export interface INEUAndReq {
    type: "and";
    values: INEUReq[];
  }
  
  /**
   * A SearchNEU OR prerequisite object.
   *
   * @param type   The type of the SearchNEU prerequisite.
   * @param values The prerequisites of which one must be completed for this
   *   prerequisite to be marked as done.
   */
  export interface INEUOrReq {
    type: "or";
    values: INEUReq[];
  }
// ...
  export type INEUReqError =
    | INEUReqCourseError
    | INEUReqAndError
    | INEUReqOrError;
  
  export enum ReqErrorType {
    COURSE = "course",
    AND = "and",
    OR = "or",
  }
  
  export interface INEUReqCourseError {
    type: ReqErrorType.COURSE;
    subject: string;
    classId: string;
  }
  
  export interface INEUReqAndError {
    type: ReqErrorType.AND;
    missing: INEUReqError[];
    subject?: string;
    classId?: string;
  }
  
  export interface INEUReqOrError {
    type: ReqErrorType.OR;
    missing: INEUReqError[];
    subject?: string;
    classId?: string;
  }
// ...
  export interface IRequiredCourse {
    type: "COURSE";
    classId: number;
    subject: string;
    description?: string;
  }
// ...
  export interface ScheduleTerm2<T> {
    season: SeasonEnum;
    status: StatusEnum;
    classes: ScheduleCourse2<T>[];
    id: T;
  }
// ...
  export interface ScheduleCourse2<T> {
    name: string;
    classId: string;
    subject: string;
    prereqs?: INEUAndReq | INEUOrReq;
    coreqs?: INEUAndReq | INEUOrReq;
    nupaths?: NUPathEnum[];
    numCreditsMin: number;
    numCreditsMax: number;
    id: T;
    generic?: boolean;
  }